Move [slug].jsx and projects index page to pages root
